[fix] URGENT: Fix exit upon ``EOF``
if you run this code:

    (load "library.scm")
    (library-print "Hello World!")

What will happen? Well, before this path, the code would load
library.scm, but then realize "This is the end of the file!",
and what it does there is that it exits. You can solve
this by some simply C++ but it was scary nonetheless.